+++ packages/viewer/web/src/getPluginViewers.jsx
@@ -10,6 +10,7 @@ import {
   normalizeUrl,
   IMAGE_TYPE,
   GALLERY_TYPE,
+  TABLE_TYPE,
 } from 'wix-rich-content-common';
 import { getBlockIndex } from './utils/draftUtils';
 import RichContentViewer from './RichContentViewer';
@@ -66,9 +67,10 @@ class PluginViewer extends PureComponent {
     return newConfig;
   };
 
-  innerRCV = ({ contentState, textAlignment, direction }) => {
+  innerRCV = ({ contentState, textAlignment, direction, renderedIn }) => {
     const { innerRCEViewerProps } = this.props;
     const config = this.removeExpand(innerRCEViewerProps.config);
+    const renderedInTable = renderedIn === TABLE_TYPE;
     return (
       <RichContentViewer
         initialState={contentState}
@@ -77,6 +79,7 @@ class PluginViewer extends PureComponent {
         {...innerRCEViewerProps}
         config={config}
         isInnerRcv
+        renderedInTable={renderedInTable}
       />
     );
   };